Reorganized testsuite
