+++ phpunit.xml
@@ -39,23 +39,24 @@
             <directory>src/adapter/etl-adapter-google-sheet/tests/Flow/ETL/Adapter/GoogleSheet/Tests/Unit</directory>
         </testsuite>
 
-        <testsuite name="integration">
+        <testsuite name="integration-io">
             <directory>src/core/etl/tests/Flow/ETL/Tests/Integration</directory>
-            <directory>src/lib/doctrine-dbal-bulk/tests/Flow/Doctrine/Bulk/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-amphp/tests/Flow/ETL/Async/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-avro/tests/Flow/ETL/Adapter/Avro/FlixTech/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-csv/tests/Flow/ETL/Adapter/CSV/Tests/Integration</directory>
-            <directory>src/adapter/etl-adapter-doctrine/tests/Flow/ETL/Adapter/Doctrine/Tests/Integration</directory>
-            <directory>src/adapter/etl-adapter-elasticsearch/tests/Flow/ETL/Adapter/Elasticsearch/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-json/tests/Flow/ETL/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-parquet/tests/Flow/ETL/Adapter/Parquet/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-reactphp/tests/Flow/ETL/Async/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-text/tests/Flow/ETL/Adapter/Text/Tests/Integration</directory>
             <directory>src/adapter/etl-adapter-xml/tests/Flow/ETL/Tests/Integration</directory>
+            <directory>src/adapter/etl-adapter-http/tests/Flow/ETL/Tests/Integration</directory>
         </testsuite>
 
-        <testsuite name="functional">
-            <directory>src/adapter/etl-adapter-http/tests/Flow/ETL/Tests/Functional</directory>
+        <testsuite name="integration-services">
+            <directory>src/core/etl/tests/Flow/ETL/Tests/Integration</directory>
+            <directory>src/lib/doctrine-dbal-bulk/tests/Flow/Doctrine/Bulk/Tests/Integration</directory>
+            <directory>src/adapter/etl-adapter-doctrine/tests/Flow/ETL/Adapter/Doctrine/Tests/Integration</directory>
+            <directory>src/adapter/etl-adapter-elasticsearch/tests/Flow/ETL/Adapter/Elasticsearch/Tests/Integration</directory>
         </testsuite>
     </testsuites>
 </phpunit>
